create common divider component
