+++ profile/src/main/kotlin/com/imashnake/animite/profile/ui/MediaTrackingList.kt
@@ -1,11 +1,5 @@
 package com.imashnake.animite.profile.ui
 
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -30,13 +24,11 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.toShape
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.snapshots.SnapshotStateMap
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.pluralStringResource
@@ -50,6 +42,7 @@ import com.imashnake.animite.api.anilist.sanitize.media.Media
 import com.imashnake.animite.api.anilist.sanitize.profile.User
 import com.imashnake.animite.api.anilist.sanitize.profile.User.ListNames.Companion.sanitize
 import com.imashnake.animite.core.ui.LocalPaddings
+import com.imashnake.animite.core.ui.component.Divider
 import com.imashnake.animite.core.ui.component.DropDownIcon
 import com.imashnake.animite.core.ui.component.MediaTrackingCard
 import com.imashnake.animite.media.MediaPage
@@ -210,7 +203,7 @@ private fun MediaTrackingItem(
                     }
 
                     if (item.season != null && item.format != null) {
-                        Divider()
+                        Divider(shape = MaterialShapes.Triangle.toShape())
                     }
 
                     item.season?.let {
@@ -224,7 +217,7 @@ private fun MediaTrackingItem(
                     }
 
                     if (item.episodes != null && item.season != null) {
-                        Divider()
+                        Divider(shape = MaterialShapes.Triangle.toShape())
                     }
 
                     item.episodes?.let {
@@ -255,31 +248,4 @@ private fun MediaTrackingItem(
             )
         }
     }
-}
-
-@OptIn(ExperimentalMaterial3ExpressiveApi::class)
-@Composable
-fun Divider(
-    modifier: Modifier = Modifier
-) {
-    val infiniteTransition = rememberInfiniteTransition(label = "divider")
-    val angle by infiniteTransition.animateFloat(
-        initialValue = 0f,
-        targetValue = 360f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(6000, easing = LinearEasing),
-            repeatMode = RepeatMode.Restart
-        ),
-        label = "rotation"
-    )
-    Box(
-        modifier = modifier
-            .graphicsLayer { rotationZ = angle }
-            .padding(horizontal = LocalPaddings.current.small)
-            .size(4.dp)
-            .background(
-                color = MaterialTheme.colorScheme.onBackground.copy(alpha = 0.74f),
-                shape = MaterialShapes.Triangle.toShape()
-            )
-    )
 }
